Set hash-table size to number of entries in hash-table-copy
As reported by John Croisant before this patch hash-table-copied
tables would report a size of 0.  This is due to *make-hash-table
setting the number of entries to 0 regardless of the size of the
hash-table's vector.

This patch also adds a test to the testsuite.

Signed-off-by: felix <felix@call-with-current-continuation.org>
(cherry picked from commit 735a6304b138dce0ca5a47577ff23a9a884d12df)

Conflicts:

	srfi-69.scm (aesthetic difference)
	tests/hash-table-tests.scm (some other tests are not backported)